refactor(detail-pages): move tabular detail page to detail pages folder
