+++ packages/application-components/src/components/detail-pages/tabular-detail-page/README.md
