+++ packages/application-components/src/index.ts
@@ -3,7 +3,6 @@ export { default as version } from './version';
 // TabHeader & Tabular pages
 export { default as TabHeader } from './components/tab-header';
 export { default as TabularMainPage } from './components/tabular-main-page';
-export { default as TabularDetailPage } from './components/tabular-detail-page';
 
 // Dialogs
 export { default as InfoDialog } from './components/dialogs/info-dialog';
@@ -20,6 +19,7 @@ export { default as CustomFormModalPage } from './components/modal-pages/custom-
 export { default as CustomFormDetailPage } from './components/detail-pages/custom-form-detail-page';
 export { default as FormDetailPage } from './components/detail-pages/form-detail-page';
 export { default as InfoDetailPage } from './components/detail-pages/info-detail-page';
+export { default as TabularDetailPage } from './components/detail-pages/tabular-detail-page';
 
 // Public page
 export { default as PublicPageLayout } from './components/public-page-layout';
